Fixed Rosenbrock potential undefined behavior

+++ include/Problem/Rosenbrock.h
@@ -16,7 +16,7 @@ public:
     double evaluate(const std::vector<double>& x) const override {
         double sum = 0.0;
         
-        for(int i = 0; i < x.size() - 1; i++) {
+        for(size_t i = 0; i < x.size() - 1; i++) {
             // Calculate inner terms
             const double term1 = x[i] * x[i] - x[i] + 1;
             const double term2 = 1 - x[i];
